+++ R/add_resource.R
@@ -33,8 +33,8 @@
 #' # Load the example Data Package
 #' package <- example_package
 #'
-#' # List the resource names
-#' package$resource_names
+#' # List resources
+#' resources(package)
 #'
 #' # Create a data frame
 #' df <- data.frame(
@@ -49,17 +49,17 @@
 #' # Add resource "positions" to the Data Package, from the data frame
 #' package <- add_resource(package, "positions", data = df)
 #'
-#' # Add resource "positions2" to the Data Package, with user-defined schema
+#' # Add resource "positions_2" to the Data Package, with user-defined schema
 #' my_schema <- create_schema(df)
-#' package <- add_resource(package, "positions2", data = df, schema = my_schema)
+#' package <- add_resource(package, "positions_2", data = df, schema = my_schema)
 #'
-#' # Add resource "observations2" to the Data Package, from CSV file paths
+#' # Add resource "observations_2" to the Data Package, from CSV file paths
 #' path_1 <- system.file("extdata", "observations_1.csv", package = "frictionless")
 #' path_2 <- system.file("extdata", "observations_2.csv", package = "frictionless")
-#' package <- add_resource(package, "observations2", data = c(path_1, path_2))
+#' package <- add_resource(package, "observations_2", data = c(path_1, path_2))
 #'
-#' # List the resource names ("positions", "positions2" & "observations2" added)
-#' package$resource_names
+#' # List resources ("positions", "positions_2", "observations_2" added)
+#' resources(package)
 add_resource <- function(package, resource_name, data, schema = NULL,
                          delim = ",") {
   # Check package
@@ -77,7 +77,7 @@ add_resource <- function(package, resource_name, data, schema = NULL,
 
   # Check resource is absent
   assertthat::assert_that(
-    !resource_name %in% package$resource_names,
+    !resource_name %in% resources(package),
     msg = glue::glue(
       "`package` already contains a resource named `{resource_name}`."
     )
@@ -146,9 +146,6 @@ add_resource <- function(package, resource_name, data, schema = NULL,
 
   # Add resource (needs to be wrapped in its own list)
   package$resources <- append(package$resources, list(resource))
-
-  # Add resource_name
-  package$resource_names <- append(package$resource_names, resource_name)
 
   package
 }